requirements.txt
owslib==0.29.2
requests==2.31.0
pandas==2.0.3
geopandas==0.13.2
Shapely==2.0.1
pytest==7.4.0
pytest-cov==4.1.0
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.18
geoalchemy2==0.14.0